--- src/main/kotlin/com/iplease/server/ip/manage/assign/data/dto/AssignedIpDto.kt
@@ -1,12 +1,0 @@
-package com.iplease.server.ip.manage.assign.data.dto
-
-import java.time.LocalDate
-
-data class AssignedIpDto(
-    val uuid: Long,
-    val issuerUuid: Long,
-    val assignerUuid: Long,
-    val assignedAt: LocalDate,
-    val expireAt: LocalDate,
-    val ip: IpDto
-)
