--- src/main/kotlin/com/iplease/server/ip/manage/assign/data/dto/IpDto.kt
@@ -1,8 +1,0 @@
-package com.iplease.server.ip.manage.assign.data.dto
-
-data class IpDto (
-    val first: Int,
-    val second: Int,
-    val third: Int,
-    val fourth: Int
-)
