--- src/main/kotlin/com/iplease/server/ip/manage/assign/exception/AlreadyExistsAssignedIpException.kt
@@ -1,3 +1,0 @@
-package com.iplease.server.ip.manage.assign.exception
-
-class AlreadyExistsAssignedIpException(val ip: String) : RuntimeException("이미 할당된 IP주소입니다! - $ip")
