--- src/main/kotlin/com/iplease/server/ip/manage/assign/exception/WrongExpireDateException.kt
@@ -1,5 +1,0 @@
-package com.iplease.server.ip.manage.assign.exception
-
-import java.time.LocalDate
-
-class WrongExpireDateException(val date: LocalDate, msg: String) : RuntimeException("잘못된 만료일 입니다! $msg - $date")
