--- src/main/kotlin/com/iplease/server/ip/manage/assign/repository/AssignedIpRepository.kt
@@ -1,9 +1,0 @@
-package com.iplease.server.ip.manage.assign.repository
-
-import com.iplease.server.ip.manage.assign.data.table.AssignedIpTable
-import org.springframework.data.r2dbc.repository.R2dbcRepository
-import reactor.core.publisher.Mono
-
-interface AssignedIpRepository: R2dbcRepository<AssignedIpTable, Long> {
-    fun existsByIpFirstAndIpSecondAndIpThirdAndIpFourth(ipFirst: Int, ipSecond: Int, ipThird: Int, ipFourth: Int): Mono<Boolean>
-}
