--- src/main/kotlin/com/iplease/server/ip/manage/assign/service/IpAssignService.kt
@@ -1,8 +1,0 @@
-package com.iplease.server.ip.manage.assign.service
-
-import com.iplease.server.ip.manage.assign.data.dto.AssignedIpDto
-import reactor.core.publisher.Mono
-
-interface IpAssignService {
-    fun assign(assignedIpDto: AssignedIpDto): Mono<AssignedIpDto>
-}
