--- src/main/kotlin/com/iplease/server/ip/manage/assign/util/DateUtil.kt
@@ -1,7 +1,0 @@
-package com.iplease.server.ip.manage.assign.util
-
-import java.time.LocalDate
-
-interface DateUtil {
-    fun dateNow(): LocalDate
-}
