+++ src/main/kotlin/com/iplease/server/ip/manage/domain/assign/util/DateUtil.kt
@@ -1,0 +1,7 @@
+package com.iplease.server.ip.manage.domain.assign.util
+
+import java.time.LocalDate
+
+interface DateUtil {
+    fun dateNow(): LocalDate
+}
