+++ src/main/kotlin/com/iplease/server/ip/manage/domain/assign/util/DateUtilImpl.kt
@@ -1,0 +1,9 @@
+package com.iplease.server.ip.manage.domain.assign.util
+
+import org.springframework.stereotype.Component
+import java.time.LocalDate
+
+@Component
+class DateUtilImpl: DateUtil {
+    override fun dateNow() = LocalDate.now()!!
+}
