--- src/main/kotlin/com/iplease/server/ip/manage/event/data/type/Error.kt
@@ -1,7 +1,0 @@
-package com.iplease.server.ip.manage.event.data.type
-
-enum class Error(
-    val routingKey: String
-) {
-    IP_ASSIGNED("v1.error.ip.demand.success"),
-}
