--- src/main/kotlin/com/iplease/server/ip/manage/event/data/type/Event.kt
@@ -1,8 +1,0 @@
-package com.iplease.server.ip.manage.event.data.type
-
-enum class Event(
-    val routingKey: String
-) {
-    IP_ASSIGNED("v1.event.ip.demand.success"),
-    IP_RELEASED("v1.event.ip.release.released")
-}
