--- src/main/kotlin/com/iplease/server/ip/manage/event/listener/EventListener.kt
@@ -1,7 +1,0 @@
-package com.iplease.server.ip.manage.event.listener
-
-import org.springframework.amqp.core.Message
-
-interface EventListener {
-    fun handle(message: Message)
-}
