--- src/main/kotlin/com/iplease/server/ip/manage/event/service/EventPublishService.kt
@@ -1,5 +1,0 @@
-package com.iplease.server.ip.manage.event.service
-
-interface EventPublishService {
-    fun <T: Any> publish(routingKey: String, data: T): T
-}
