--- src/main/kotlin/com/iplease/server/ip/manage/event/service/EventSubscribeService.kt
@@ -1,7 +1,0 @@
-package com.iplease.server.ip.manage.event.service
-
-import com.iplease.server.ip.manage.event.listener.EventListener
-
-interface EventSubscribeService {
-    fun addListener(listener: EventListener)
-}
