+++ src/main/kotlin/com/iplease/server/ip/manage/global/assign/data/dto/IpDto.kt
@@ -1,0 +1,8 @@
+package com.iplease.server.ip.manage.global.assign.data.dto
+
+data class IpDto (
+    val first: Int,
+    val second: Int,
+    val third: Int,
+    val fourth: Int
+)
