+++ src/main/kotlin/com/iplease/server/ip/manage/global/assign/data/table/AssignedIpTable.kt
@@ -1,0 +1,19 @@
+package com.iplease.server.ip.manage.global.assign.data.table
+
+import org.springframework.data.annotation.Id
+import org.springframework.data.relational.core.mapping.Table
+import java.time.LocalDate
+
+@Table
+data class AssignedIpTable (
+    @Id
+    val uuid: Long,
+    val issuerUuid: Long,
+    val assignerUuid: Long,
+    val assignedAt: LocalDate,
+    val expireAt: LocalDate,
+    val ipFirst: Int,
+    val ipSecond: Int,
+    val ipThird: Int,
+    val ipFourth: Int
+)
