+++ src/main/kotlin/com/iplease/server/ip/manage/global/assign/repository/AssignedIpRepository.kt
@@ -1,0 +1,9 @@
+package com.iplease.server.ip.manage.global.assign.repository
+
+import com.iplease.server.ip.manage.global.assign.data.table.AssignedIpTable
+import org.springframework.data.r2dbc.repository.R2dbcRepository
+import reactor.core.publisher.Mono
+
+interface AssignedIpRepository: R2dbcRepository<AssignedIpTable, Long> {
+    fun existsByIpFirstAndIpSecondAndIpThirdAndIpFourth(ipFirst: Int, ipSecond: Int, ipThird: Int, ipFourth: Int): Mono<Boolean>
+}
