+++ src/main/kotlin/com/iplease/server/ip/manage/infra/event/data/type/Error.kt
@@ -1,0 +1,7 @@
+package com.iplease.server.ip.manage.infra.event.data.type
+
+enum class Error(
+    val routingKey: String
+) {
+    IP_ASSIGNED("v1.error.ip.demand.success"),
+}
