+++ src/main/kotlin/com/iplease/server/ip/manage/infra/event/listener/EventListener.kt
@@ -1,0 +1,7 @@
+package com.iplease.server.ip.manage.infra.event.listener
+
+import org.springframework.amqp.core.Message
+
+interface EventListener {
+    fun handle(message: Message)
+}
