+++ src/main/kotlin/com/iplease/server/ip/manage/infra/event/service/EventPublishService.kt
@@ -1,0 +1,5 @@
+package com.iplease.server.ip.manage.infra.event.service
+
+interface EventPublishService {
+    fun <T: Any> publish(routingKey: String, data: T): T
+}
