+++ src/main/kotlin/com/iplease/server/ip/manage/infra/event/service/EventSubscribeService.kt
@@ -1,0 +1,7 @@
+package com.iplease.server.ip.manage.infra.event.service
+
+import com.iplease.server.ip.manage.infra.event.listener.EventListener
+
+interface EventSubscribeService {
+    fun addListener(listener: EventListener)
+}
