+++ src/main/kotlin/com/iplease/server/ip/manage/infra/log/type/LoggerType.kt
@@ -1,0 +1,8 @@
+package com.iplease.server.ip.manage.infra.log.type
+
+enum class LoggerType {
+    EVENT_PUBLISH_LOGGER,
+    EVENT_SUBSCRIBE_LOGGER,
+    IP_ASSIGN_LOGGER
+    ;
+}
