+++ src/main/kotlin/com/iplease/server/ip/manage/infra/log/util/LoggerUtil.kt
@@ -1,0 +1,7 @@
+package com.iplease.server.ip.manage.infra.log.util
+
+interface LoggerUtil<IN, OUT> {
+    fun logOnStart(input: IN, uuid: String)
+    fun logOnComplete(output: OUT, uuid: String)
+    fun logOnError(throwable: Throwable, uuid: String)
+}
