--- src/main/kotlin/com/iplease/server/ip/manage/log/type/LoggerType.kt
@@ -1,8 +1,0 @@
-package com.iplease.server.ip.manage.log.type
-
-enum class LoggerType {
-    EVENT_PUBLISH_LOGGER,
-    EVENT_SUBSCRIBE_LOGGER,
-    IP_ASSIGN_LOGGER
-    ;
-}
