--- src/main/kotlin/com/iplease/server/ip/manage/log/util/LoggerUtil.kt
@@ -1,7 +1,0 @@
-package com.iplease.server.ip.manage.log.util
-
-interface LoggerUtil<IN, OUT> {
-    fun logOnStart(input: IN, uuid: String)
-    fun logOnComplete(output: OUT, uuid: String)
-    fun logOnError(throwable: Throwable, uuid: String)
-}
